레이어에 맞춰 패키지를 이동하라
