+++ src/main/java/com/realestate/service/property/CreatePropertyUseCase.java
@@ -1,6 +1,5 @@
 package com.realestate.service.property;
 
-import com.realestate.service.property.dto.CreatePropertyCommand;
 import com.realestate.service.property.entity.Property;
 
 /**
