+++ src/main/java/com/realestate/service/property/UpdatePropertyCommand.java
@@ -1,30 +1,27 @@
-package com.realestate.service.property.dto;
+package com.realestate.service.property;
 
 import java.time.LocalDate;
 
-import com.realestate.service.property.address.entity.PropertyAddress;
 import com.realestate.service.property.constant.ContractType;
 import com.realestate.service.property.constant.ResidentialType;
 import com.realestate.service.property.constant.StructureType;
-import com.realestate.service.property.entity.Property;
 import com.realestate.service.property.entity.PropertyFloor;
 import com.realestate.service.property.entity.PropertyInformation;
 import com.realestate.service.property.entity.PropertyPrice;
-import com.realestate.service.user.entity.User;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class CreatePropertyCommand {
+public class UpdatePropertyCommand {
 
   private final long userId;
   private final PropertyInformationCommand propertyInformationCommand;
   private final PropertyAddressCommand propertyAddressCommand;
 
   /**
-   * CreatePropertyCommand 를 생성하여 리턴합니다.
+   * UpdatePropertyCommand 를 생성하여 리턴합니다.
    */
-  public CreatePropertyCommand(long userId,
+  public UpdatePropertyCommand(long userId,
                                PropertyInformationCommand propertyInformationCommand,
                                PropertyAddressCommand propertyAddressCommand) {
     this.userId = userId;
@@ -115,54 +112,29 @@ public class CreatePropertyCommand {
     }
   }
 
-
   /**
-   * 매물 주소 엔티티를 반환합니다.
+   * 매물 정보를 생성 후 반환합니다.
    */
-  public PropertyAddress toAddressEntity(Property property) {
-    return PropertyAddress.builder()
-        .property(property)
-        .city(propertyAddressCommand.getCity())
-        .address(propertyAddressCommand.getAddress())
-        .roadAddress(propertyAddressCommand.getRoadAddress())
-        .zipcode(propertyAddressCommand.getZipcode())
-        .latitude(propertyAddressCommand.getLatitude())
-        .longitude(propertyAddressCommand.getLongitude())
-        .build();
-  }
-
-  /**
-   * 매물 엔티티를 반환합니다.
-   */
-  public Property toEntity(User user) {
-    return Property.builder()
-        .user(user)
-        .title(propertyInformationCommand.getTitle())
-        .content(propertyInformationCommand.getContent())
-        .propertyInformation(createPropertyInformation())
-        .build();
-  }
-
-  private PropertyInformation createPropertyInformation() {
+  public PropertyInformation createPropertyInformation() {
     return PropertyInformation.builder()
-        .propertyPrice(createPropertyPrice())
-        .propertyFloor(createPropertyFloor())
+        .area(propertyInformationCommand.getArea())
         .structureType(propertyInformationCommand.getStructureType())
         .contractType(propertyInformationCommand.getContractType())
+        .residentialType(propertyInformationCommand.getResidentialType())
         .availableParking(propertyInformationCommand.getAvailableParking())
         .moveInDate(propertyInformationCommand.getMoveInDate())
         .completionDate(propertyInformationCommand.getCompletionDate())
-        .area(propertyInformationCommand.getArea())
-        .residentialType(propertyInformationCommand.getResidentialType())
+        .propertyPrice(createPropertyPrice())
+        .propertyFloor(createPropertyFloor())
         .build();
   }
 
   private PropertyPrice createPropertyPrice() {
     return PropertyPrice.builder()
+        .sellPrice(propertyInformationCommand.getSellPrice())
+        .deposit(propertyInformationCommand.getDeposit())
         .adminPrice(propertyInformationCommand.getAdminPrice())
         .monthlyPrice(propertyInformationCommand.getMonthlyPrice())
-        .deposit(propertyInformationCommand.getDeposit())
-        .sellPrice(propertyInformationCommand.getSellPrice())
         .build();
   }
