+++ src/main/java/com/realestate/service/property/UpdatePropertyUseCase.java
@@ -1,6 +1,5 @@
 package com.realestate.service.property;
 
-import com.realestate.service.property.dto.UpdatePropertyCommand;
 import com.realestate.service.property.entity.Property;
 
 /**
