+++ src/test/java/com/realestate/service/property/UpdatePropertyServiceTest.java
@@ -12,7 +12,6 @@ import com.realestate.service.property.address.entity.PropertyAddressRepository;
 import com.realestate.service.property.constant.ContractType;
 import com.realestate.service.property.constant.ResidentialType;
 import com.realestate.service.property.constant.StructureType;
-import com.realestate.service.property.dto.UpdatePropertyCommand;
 import com.realestate.service.property.entity.Property;
 import com.realestate.service.property.entity.PropertyRepository;
 import org.junit.jupiter.api.BeforeEach;
